+++ src/main/lombok/de/larssh/json/dom/JsonDomDocument.java
@@ -15,6 +15,7 @@ import org.w3c.dom.Node;
 import org.w3c.dom.ProcessingInstruction;
 import org.w3c.dom.Text;
 
+import de.larssh.json.dom.values.JsonDomValue;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import lombok.EqualsAndHashCode;