+++ src/main/lombok/de/larssh/json/dom/children/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * JSON DOM children
+ */
+@de.larssh.utils.annotations.NonNullByDefault
+package de.larssh.json.dom.children;