+++ src/main/lombok/de/larssh/json/dom/values/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * JSON DOM values
+ *
+ * <p>
+ * JSON DOM does not come with its own JSON parser or JSON element objects.
+ * Instead {@link de.larssh.json.dom.values.JsonDomValue} is made up to
+ * integrate existing parsers.
+ */
+@de.larssh.utils.annotations.NonNullByDefault
+package de.larssh.json.dom.values;